+++ src/lib/server/db/migrations/0001_swift_lionheart.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "entry" ADD COLUMN "description" varchar(1024);