+++ src/routes/(app)/entries/create/_components/FilesSection.svelte
@@ -6,14 +6,15 @@
 	import { getContext } from 'svelte';
 	import type { SuperForm } from 'sveltekit-superforms';
 	import SubmittingTo from './SubmittingTo.svelte';
-	// TODO Add proof field
 	let { form } = getContext<SuperForm<CreateEntrySchema>>('superform');
 	let { isValid = $bindable() }: { isValid: boolean } = $props();
 	$effect(() => {
 		isValid = createEntrySchema
 			.pick({
 				mediaId: true,
-				thumbnailId: true // proofId: true
+				thumbnailId: true,
+				previewId: true,
+				proofId: true
 			})
 			.safeParse($form).success;
 	});
@@ -33,13 +34,16 @@
 			labelProps={{ class: 'text-2xl font-bold' }}
 		/>
 	</div>
-	<!-- <FileField
-		mediaType="archive"
-		maxFileSize={1024 * 1024 * 2}
-		field="proofId"
-		label="Proof"
-		labelProps={{ class: 'text-2xl font-bold' }}
-	/> -->
+	<div class="mt-sm w-full">
+		<FileField
+			customUploadUrl="/api/uploads/proof"
+			mediaType="archive"
+			maxFileSize={1024 * 1024 * 512}
+			field="proofId"
+			label="Proof"
+			labelProps={{ class: 'text-2xl font-bold' }}
+		/>
+	</div>
 	<Divider class="my-xl" />
 	<div class="gap-sm flex flex-col">
 		<h3>What are the different files for?</h3>
@@ -54,10 +58,10 @@
 			This is the image that will be portrayed on the website for viewers to enjoy and vote on. Make
 			it pop, but don’t clickbait...
 		</p>
-		<!-- <h4>Proof</h4>
+		<h4>Proof</h4>
 		<p class="text-shade-300">
 			If the category requires it, you will upload a compressed file with suggested proof of
 			authorship here.
-		</p> -->
+		</p>
 	</div>
 </section>