+++ src/routes/(app)/entries/create/_components/FinishSection.svelte
@@ -2,6 +2,7 @@
 	import SelectField from '$lib/components/Form/SelectField.svelte';
 	import type { StyledSelectItem } from '$lib/components/Form/StyledSelect.svelte';
 	import TextField from '$lib/components/Form/TextField.svelte';
+	import TextAreaField from '$lib/components/Form/TextAreaField.svelte';
 	import Link from '$lib/components/Link.svelte';
 	import { licenseCode } from '$lib/schemas/license';
 	import { createEntrySchema, type CreateEntrySchema } from '$lib/schemas/entry';
@@ -9,6 +10,7 @@
 	import type { SuperForm } from 'sveltekit-superforms';
 	import SubmittingTo from './SubmittingTo.svelte';
 	import Divider from '$lib/components/Divider.svelte';
+	import { t } from '$lib/i18n';
 	let { isValid = $bindable() }: { isValid: boolean } = $props();
 	const licenses: StyledSelectItem[] = licenseCode.map((code) => ({
 		label: code.toUpperCase(),
@@ -26,21 +28,36 @@
 	<SubmittingTo />
 	<Divider class="my-xl" />
 	<div class="gap-xs flex flex-col">
-		<TextField field="title" label="Title" labelProps={{ class: 'text-2xl font-bold' }} />
-		<p class="text-shade-300">What do you call your masterpiece?</p>
-	</div>
-	<div class="gap-xs flex flex-col">
-		<SelectField
-			type="single"
-			field="license"
-			label="License"
-			labelProps={{ class: 'text-2xl font-bold' }}
-			items={licenses}
-		/>
-		<div class="text-shade-300">
-			Confused? Use <Link href="https://chooser-beta.creativecommons.org" target="_blank">
-				this tool
-			</Link> to help you pick a license.
+		<div class="gap-xs flex flex-col">
+			<TextField field="title" label="Title" labelProps={{ class: 'text-2xl font-bold' }} />
+			<p class="text-shade-300">{$t('What do you call your masterpiece?')}</p>
+		</div>
+		<div class="gap-xs flex flex-col">
+			<TextAreaField
+				field="description"
+				label="Description"
+				labelProps={{ class: 'text-2xl font-bold' }}
+				rows={3}
+			/>
+			<p class="text-shade-300">
+				{$t('Describe your creative process, give a shoutout to your mom or bribe the voters.')}
+			</p>
+		</div>
+		<div class="gap-xs flex flex-col">
+			<SelectField
+				type="single"
+				field="license"
+				label="License"
+				labelProps={{ class: 'text-2xl font-bold' }}
+				items={licenses}
+			/>
+			<div class="text-shade-300">
+				{$t('Confused? Use')}
+				<Link href="https://chooser-beta.creativecommons.org" target="_blank">
+					{$t('this tool')}
+				</Link>
+				{$t('to help you pick a license.')}
+			</div>
 		</div>
 	</div>
 </section>