+++ CHANGELOG.md
@@ -76,6 +76,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [#103](https://github.com/thejoltjoker/kreacon/pull/103)
 - Add email verification
   [#127](https://github.com/thejoltjoker/kreacon/pull/127)
+- Setup Azure Storage containers
+  [#128](https://github.com/thejoltjoker/kreacon/pull/128)
+- Add proof file and description field to entries
+  [#136](https://github.com/thejoltjoker/kreacon/pull/136)
+- Add file validation to drop zone
+  [#138](https://github.com/thejoltjoker/kreacon/pull/138)
 
 ### Changed